There's a bug report at
http://blackfin.uclinux.org/gf/project/toolchain/forum/?action=ForumBrowse&forum_id=44&_forum_action=ForumMessageBrowse&thread_id=17464

indicating that libpthread does not behave correctly when a thread is created
with a stack set up by the user in the data section.  On nommu systems, we
need to update the estimate for the bottom of the initial thread's stack
whenever we create a new thread.  Currently, this is only done when the user
does not pass in a stack he has set up himself.  Fixed by copying some code
to the other branch of the if statement.



git-svn-id: svn://localhost/svn/toolchain/trunk@1642 2adf4935-75e9-4407-93b2-34b21a1880a4