Upstream r18536 | vapier | 2007-05-02 10:05:09 +0200 (Wed, 02 May 2007) | 7 lines
Nickolai Zeldovich writes:
Currently, tcgetpgrp() in uClibc uses an int to store a PID (fetched
via ioctl TIOCGPGRP).  This causes problems on platforms where pid_t
is defined to be larger (e.g., uint64_t).  Other functions in termios,
such as tcgetsid() and tcsetpgrp(), already pass a pid_t to ioctl(),
so the following patch does the same in tcgetpgrp() as well.



git-svn-id: svn://localhost/svn/toolchain/trunk@1945 2adf4935-75e9-4407-93b2-34b21a1880a4