Upstream r19152 | sjhill | 2007-07-18 05:16:05 +0200 (Wed, 18 Jul 2007) | 2 lines
Enable all of the remaining baud rates. All of the other architectures have them enabled, so let's go all the way. There was a quick discussion on IRC about making the max baud rate configurable. That requires a bunch of #ifdef salting throughout all of the 'termios.h' files for the pre-processor to chew on. I'm going to pass on that for now.

[Modified, since we already had a patch for this in our tree, but it didn't
match upstream.]


git-svn-id: svn://localhost/svn/toolchain/trunk@1972 2adf4935-75e9-4407-93b2-34b21a1880a4