Upstream r20318 | jocke | 2007-10-23 13:47:01 +0200 (Tue, 23 Oct 2007) | 10 lines
Poll with zero timeout

Jean-Christian de Rivaz writes:
The attached patch solve an issue I faced while using the libdbus-glib
waiting for a D-Bus message or the end of a glib timer at the same time.
This specific case of use generate a poll call with a zero timeout. On
platformes with the glibc a zero timeout poll return immetiately even if
there is no file descriptor event. But on platformes with uClibc a zero
timeout poll block until a file descriptor event occurs.

 -- I think this is our bug [#3590]


git-svn-id: svn://localhost/svn/toolchain/trunk@2006 2adf4935-75e9-4407-93b2-34b21a1880a4