------------------------------------------------------------------------ r20712 | vapier | 2008-01-05 06:55:23 +0100 (Sat, 05 Jan 2008) | 8 lines
michael_d writes in #1874: 
On an i386 platform with no rt_sigsuspend syscall (ie: Linux 2.0), compilation will halt on libc/sysdeps/linux/common/sigsuspend.os with a cryptic error message:
"Error: non-constant expression in ".if" statement"

I've investigated and found that the cause is that a literal '0' is being passed into a block of complex assembler macrology that is only prepared to deal with register names - '%eax', etc.

In turn, that seems to be because of a typo in the GCC register constraints. The constraints for 2 and 3-argument syscalls includes a "C" constraint. To gcc, "C" means an SSE floating point constant -- an unlikely element in a syscall. I suspect the author meant to type "S" (%esi).

------------------------------------------------------------------------


git-svn-id: svn://localhost/svn/toolchain/trunk@2423 2adf4935-75e9-4407-93b2-34b21a1880a4