------------------------------------------------------------------------ r20714 | vapier | 2008-01-05 07:07:49 +0100 (Sat, 05 Jan 2008) | 5 lines
michael_d writes in 1879:
When no TIOCGPTN definition is present in the kernel headers, the library's ptsname() function will not work.

The libc/stdlib/ptsname_r.c file is the problem. This file includes a complicated nest of #if directives. One of these #if's has the opposite sense from what is required.

------------------------------------------------------------------------


git-svn-id: svn://localhost/svn/toolchain/trunk@2425 2adf4935-75e9-4407-93b2-34b21a1880a4