------------------------------------------------------------------------ r20907 | sjhill | 2008-01-27 00:13:37 +0100 (Sun, 27 Jan 2008) | 10 lines
Filter out '-std=gnu99' when running the assembler. This causes an error 
when building MIPS that looks like this:

libc/string/mips/memcpy.S:156:1: pasting "memcpy" and ":" does not give
a valid preprocessing token
make[1]: *** [libc/string/mips/memcpy.os] Error 1

Even if this does not cause an error on other architectures, this option
should not be having an effect on preprocessed assembly code.

------------------------------------------------------------------------


git-svn-id: svn://localhost/svn/toolchain/trunk@2473 2adf4935-75e9-4407-93b2-34b21a1880a4