------------------------------------------------------------------------ r22155 | bernds | 2008-05-31 23:39:13 +0200 (Sat, 31 May 2008) | 3 lines
Fix a dependency to mention libc.so, not libc.so.0, as we only have a rule to
build the former.

------------------------------------------------------------------------
r22161 | bernds | 2008-06-01 18:49:23 +0200 (Sun, 01 Jun 2008) | 8 lines

A slight improvement over my previous commit which ensured we always rebuild
libc when one of the source files changes.

Since there are more places which want to depend on $(libc), fix them all to
use a new variable $(libc.depend), which contains the filename for which we
have a rule.


------------------------------------------------------------------------


git-svn-id: svn://localhost/svn/toolchain/trunk@2587 2adf4935-75e9-4407-93b2-34b21a1880a4