BuildToolChain: build uClibc with relevant target rather than all with bfin-elf
the most important aspect is that it fixes [#4976] which has multiple
practical implications:
 - no longer bleeds libgcc_s.so symbols into uClibc which ...
 - avoids collides with libgcc_s.so at runtime
 - avoids wasting space with multiple libgcc_s.so copies on disk and runtime
 - fixes EXEC_STACK markings on the main C libraries

as to the change itself, the old way the toolchain script worked:
 - build entire bfin-elf
 - use bfin-elf to build all uClibc multilibs (FLAT/FDPIC)
 - build entire bfin-uclinux/bfin-linux-uclibc against the bfin-elf
   compiled uClibc

while the new way does:
 - build entire bfin-elf
 - build bfin-uclinux binutils
 - install uClibc headers only for bfin-uclinux
 - build a C-only "bare" gcc compiler for bfin-uclinux
 - build all FLAT multilibs uClibc with this "bare" bfin-uclinux compiler
 - build a normal "full" gcc compiler for bfin-uclinux
 - <do same process for bfin-linux-uclibc as bfin-uclinux>

this does mean we compile one gcc version twice per target, but the
"bare" toolchain should be much quicker as it is C-only with no shared
libs.  in my tests, it seems to add about 4min per target to the whole
build process.


git-svn-id: svn://localhost/svn/toolchain/trunk@5018 2adf4935-75e9-4407-93b2-34b21a1880a4