opcodes: blackfin: fix decoding of dsp mult insns
When assigning to a register half, the mac0 part of the mult insn
was not decoding properly.  It would always show a full dreg instead
of the dreg low half.

Once we fix the disassembler, we have to update a few of the gas
tests as their previous expected output was incorrect.


git-svn-id: svn://localhost/svn/toolchain/trunk@5235 2adf4935-75e9-4407-93b2-34b21a1880a4