ld-elf2flt: do not alias -v to --verbose
The GNU linker uses -v as a shortcut to --version, not --verbose.  So atm,
if you run `ld -v` to get the linker version, ld-elf2flt throws out a lot
of verbose debugging information.  So drop the -v checking in ld-elf2flt
to keep from breaking systems that parse the linker version.


git-svn-id: svn://localhost/svn/toolchain/trunk@5369 2adf4935-75e9-4407-93b2-34b21a1880a4