[#6611] toolchain-regtest: load gdb-comm for jtag targets
the previous boards files would first load_generic_config for "bfin" which in turn would load "gdb-comm".  we need the former first for gdb so that it will detect a supported target and we need the latter to load up the support routines to actually connect to gdbproxy for gcc.

git-svn-id: svn://localhost/svn/toolchain/trunk@5451 2adf4935-75e9-4407-93b2-34b21a1880a4