uClibc: daemon: fix up INTERNAL_SYSCALL() usage
Make sure we declare the error properly in case a port uses it, and fix the
invocation of exit().  Since clone() will be returning a pid, assume that
the value will not have the MSB set (negative) to simplify the error test a
bit more.  If gcc supports it, force this function to always be heavily
optimized in a bid to avoid stack usage as much as possible.

(this is already in upstream uClibc)


git-svn-id: svn://localhost/svn/toolchain/trunk@5790 2adf4935-75e9-4407-93b2-34b21a1880a4