gcc-4.5: fix linking of ppl against local copy of gmp
the ppl code is not passed the right configure flags, so it ends up using the system copy of gmp.  when that version is newer (like gmp-5.1.x), the whole build blows up.

we also have to pass in -lstdc++ since the local copies of gmpxx are static only and gcc tries to link some C apps against it.

git-svn-id: svn://localhost/svn/toolchain/trunk@5987 2adf4935-75e9-4407-93b2-34b21a1880a4